% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{dist_comb}
\alias{dist_comb}
\title{Two by Two Combinations of Distributions}
\usage{
dist_comb(n, samples_distributions)
}
\arguments{
\item{n}{sample size}

\item{samples_distributions}{dataframe of the distributions}
}
\description{
This function takes a data frame with different columns of samples with different distributions
and makes the two by to combination of the samples.
}
\details{
The data frame should have a specific format. The first column should be sample size and the consequent 
columns should have samples of different distributions. The output has four columns: name of the first distribution, 
name of the second distribution, the first distribution samples, and the second distribution samples.
}
