% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_rejection_rate.R
\name{ks_rejection_rate}
\alias{ks_rejection_rate}
\title{Kolmogorov Smirnov Rejection Rate}
\usage{
ks_rejection_rate(X1, X2, alpha = 0.05)
}
\arguments{
\item{X1}{matrix of the first distribution}

\item{X2}{matrix of the second distribution.}

\item{alpha}{significance level (defaul is 0.05)}
}
\description{
This function finds the rejection rate of Kolmogorov Smirnov test for two distributions.
}
\details{
The inputs of the function is two matrices of samples of two different distributions. 
Number of rows if the sample size and number of columns is 1000 which is the number of generated samples
from the distributions. The output is the number of times the test rejects the null hypothesis.
}
\examples{

# 1000 samples of size 10 generated from 
# a normal distribution with mean of 0 and sd of 1
X1 <- rnorm(1000*10, mean = 0, sd = 1) \%>\%
 matrix(ncol = 1000, nrow = 10)

# 1000 samples of size 10 generated from 
# a exponential distribution with rate of 1

X2 = rexp(1000*10, rate = 1) \%>\% 
matrix(ncol = 1000, nrow = 10)

rejection_rate <- ks_rejection_rate(X1 = X1, 
X2 = X2, alpha = 0.05)

}
