% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_anydistribution.R
\name{sample_anyDistribution}
\alias{sample_anyDistribution}
\title{Taking Samples of Size n for the Built-in R Density Functions}
\usage{
sample_anyDistribution(n, sample_fun = sample_fun, ...)
}
\arguments{
\item{n}{sample size}

\item{sample_fun}{any built-in R density function that can take a random sample of a specified distribution}

\item{...}{parameters corresponding to the specified distribution}
}
\description{
This function returns 1000 samples of size in using the built-in
density functions in R. The output is a matrix with 1000 columns and n rows.
}
\examples{

# taking samples of size 10 from a normal distribution 
# with mean 0 and sd 1

sample_anyDistribution(n = 10, sample_fun = rnorm, mean = 0, sd = 1)
}
